envs/style_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-style-env"
version = "0.1.0"
description = "Style Consistency Environment for OpenEnv - RL environment for training frontend style adherence"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git@main",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "style_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["style_env", "style_env.server", "style_env.server.scorers"]
package-dir = { "style_env" = ".", "style_env.server" = "server", "style_env.server.scorers" = "server/scorers" }

